% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparboot.R
\name{bootstrap_f_sample}
\alias{bootstrap_f_sample}
\title{Internal Bootstrap Sampling Function for F-tests}
\usage{
bootstrap_f_sample(x_val, y_val, grp_val, grp_sizes, pre_calc, na_rm)
}
\arguments{
\item{x_val}{A numeric vector of values from the primary variable.}

\item{y_val}{A numeric vector of values from the second variable. This
parameter is not used in this function but is included for consistency with
other bootstrap sampling functions.}

\item{grp_val}{A factor vector of group labels.}

\item{grp_sizes}{A table of group sizes.}

\item{pre_calc}{A list containing pre-calculated statistics from the original
data.}

\item{na_rm}{Set na.rm in places where the original algorithms do not.
Default is FALSE.}
}
\value{
A numeric vector of length 2 containing the bootstrapped test statistic
and the effect size.
}
\description{
This is an internal function used by nonparboot() to perform bootstrap
resampling for F-tests. It is not intended to be called directly by the user.
}
\keyword{internal}
