% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparboot.R
\name{bootstrap_t_sample}
\alias{bootstrap_t_sample}
\title{Internal Bootstrap Sampling Function for T-tests}
\usage{
bootstrap_t_sample(x_val, y_val, grp_val, grp_sizes, pre_calc)
}
\arguments{
\item{x_val}{A numeric vector of values from the primary variable.}

\item{y_val}{A numeric vector of values from the second variable
(only used for paired t-tests).}

\item{grp_val}{A factor vector of group labels (only used for independent
t-tests and F-test).}

\item{grp_sizes}{A table of group sizes (only used for independent t-tests and
F-test).}

\item{pre_calc}{A list containing pre-calculated statistics from the original
data.}
}
\value{
A numeric vector of length 2 containing the bootstrapped test statistic
and mean difference.
}
\description{
This is an internal function used by nonparboot() to perform bootstrap
resampling for independent t-tests.
}
\keyword{internal}
