% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_f_id}
\alias{data_f_id}
\title{Simulated Data for F-test Using Identically Distributed Data}
\format{
\subsection{\code{data_f_id}}{

A data frame with 15 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{grp}{A factor variable with group labels}
}
}
}
\source{
Simulated data
}
\usage{
data_f_id
}
\description{
A simulated data set to experiment with nonparboot() with test = "F" when all
groups are drawn from identical distributions
}
\keyword{datasets}
