% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_f_mi}
\alias{data_f_mi}
\title{Simulated Data for F-test with Missing Outcomes}
\format{
\subsection{\code{data_f_mi}}{

A data frame with 15 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{grp}{A factor variable with group labels}
}
}
}
\source{
Simulated data
}
\usage{
data_f_mi
}
\description{
A simulated data set to experiment with nonparboot() with test = "F" and
missing outcomes
}
\keyword{datasets}
