% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pt}
\alias{data_pt}
\title{Simulated Data for Paired T-test}
\format{
\subsection{\code{data_pt}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{y}{A numeric variable}
}
}
}
\source{
Simulated data
}
\usage{
data_pt
}
\description{
A simulated data set to experiment with nonparboot() with test = "pt"
}
\keyword{datasets}
