% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pt_id}
\alias{data_pt_id}
\title{Simulated Data for Paired T-test Using Identically Distributed Data}
\format{
\subsection{\code{data_pt_id}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{y}{A numeric variable}
}
}
}
\source{
Simulated data
}
\usage{
data_pt_id
}
\description{
A simulated data set to experiment with nonparboot() with test = "pt" when
both variables are drawn from identical distributions
}
\keyword{datasets}
