% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pt_mi}
\alias{data_pt_mi}
\title{Simulated Data for Paired T-test with Missing Data}
\format{
\subsection{\code{data_pt_mi}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{y}{A numeric variable}
}
}
}
\source{
Simulated data
}
\usage{
data_pt_mi
}
\description{
A simulated data set to experiment with nonparboot() with test = "pt" and
missing values
}
\keyword{datasets}
