% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t}
\alias{data_t}
\title{Simulated Data for Independent T-test}
\format{
\subsection{\code{data_t}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{grp}{A factor variable with group labels}
}
}
}
\source{
Simulated data
}
\usage{
data_t
}
\description{
A simulated data set to experiment with nonparboot() with test = "t"
}
\keyword{datasets}
