% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_t_mi}
\alias{data_t_mi}
\title{Simulated Data for Independent T-test with Missing Outcomes}
\format{
\subsection{\code{data_t_mi}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{A numeric variable}
\item{grp}{A factor variable with group labels}
}
}
}
\source{
Simulated data
}
\usage{
data_t_mi
}
\description{
A simulated data set to experiment with nonparboot() with test = "t" and
missing outcome values
}
\keyword{datasets}
