% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_helpers.R
\name{format_citation}
\alias{format_citation}
\title{Format Citation}
\usage{
format_citation(cit)
}
\arguments{
\item{cit}{A citation object typically obtained from \code{citation()}.}
}
\value{
A character string with the formatted citation.
}
\description{
This internal function formats a citation object into a readable string.
The function extracts relevant information such as the title, author,
year, address, note, and URL from the citation object and formats it into a
standardized citation format.
}
\keyword{internal}
