% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_helpers.R
\name{getCellBlocks}
\alias{getCellBlocks}
\title{Generate a List of Available Cell Blocks}
\usage{
getCellBlocks()
}
\value{
A named list where each name is a string describing a cell block
(e.g., "T2 Cell Block 1.1") and each value is a string corresponding
to the function name (e.g., "replext_t2_c1.1") that is associated with
the simulation process for that particular cell block.
}
\description{
This function creates and returns a named list of cell blocks,
where each name corresponds to a descriptive label of the cell block,
and the value is the function name associated with that cell block.
}
\keyword{internal}
