% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_helpers.R
\name{getUIParams}
\alias{getUIParams}
\title{Generate UI Elements for Selected Cell Block}
\usage{
getUIParams(cellBlock)
}
\arguments{
\item{cellBlock}{A character string identifying the selected cell block.
The function uses this parameter to determine which set of UI elements to generate.}
}
\value{
A list of Shiny UI elements specific to the selected cell block.
These UI elements include numeric inputs, text inputs, and other relevant controls
required to capture user inputs for simulation parameters.
}
\description{
This function generates a dynamic user interface (UI) for the Shiny app based on the selected cell block.
It creates a list of Shiny UI elements, such as numeric inputs and text inputs,
tailored to the requirements of the chosen cell block.
}
\keyword{internal}
