% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{nonparboot_app}
\alias{nonparboot_app}
\title{Shiny App for Nonparametric Bootstrap Tests with Pooled Resampling}
\usage{
nonparboot_app()
}
\value{
An interactive Shiny app.
}
\description{
This function creates a Shiny app for performing nonparametric bootstrap tests
with pooled resampling. The app allows you to conduct an independent t-test,
a paired t-test, or a one-way ANOVA, depending on your input.
}
\examples{
if(interactive()){
  nonparboot_app()
}

}
\references{
Dwivedi AK, Mallawaarachchi I, Alvarado LA (2017). 'Analysis of small sample
size studies using nonparametric bootstrap test with pooled resampling method.'
Statistics in Medicine, 36 (14), 2187-2205. https://doi.org/10.1002/sim.7263
}
