% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replext_helpers.R
\name{runSimulation}
\alias{runSimulation}
\title{Execute Simulation Based on User Inputs}
\usage{
runSimulation(input)
}
\arguments{
\item{input}{A list of inputs gathered from the Shiny app's UI, including the selected
cell block and other parameters necessary for the simulation.}
}
\value{
The result of the simulation function that corresponds to the selected cell block.
This result is typically a data frame containing the outcomes of the simulation.
}
\description{
This internal function manages the simulation process in a Shiny app environment.
It dynamically selects the appropriate simulation function based on the selected cell block
and passes user inputs to this function. The function also handles the setting of a random
number seed, if provided, to ensure reproducibility of results.
}
\keyword{internal}
