% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2_replext.R
\name{t2_replext_c1.1}
\alias{t2_replext_c1.1}
\title{Replicate and Extend Simulation Results from Table 2 Cell 1.1}
\usage{
t2_replext_c1.1(
  M1 = 5,
  S1 = 1,
  M2 = 5,
  S2 = 1,
  Sk1 = NULL,
  Sk2 = NULL,
  n1 = c(3, 4, 5, 6, 7, 8, 9, 10, 15),
  n2 = c(3, 4, 5, 6, 7, 8, 9, 10, 15),
  n_simulations = 10000,
  nboot = 1000,
  conf.level = 0.95
)
}
\arguments{
\item{M1}{Mean for the first group, default is 5.}

\item{S1}{Standard deviation for the first group, default is 1.}

\item{M2}{Mean for the second group, default is 5.}

\item{S2}{Standard deviation for the second group, default is 1.}

\item{Sk1}{Skewness parameter for the first group, default is NULL (normal distribution).}

\item{Sk2}{Skewness parameter for the second group, default is NULL (normal distribution).}

\item{n1}{Vector of sample sizes for the first group.}

\item{n2}{Vector of sample sizes for the second group, must be the same length as n1.}

\item{n_simulations}{Number of simulations to run, default is 10,000.}

\item{nboot}{Number of bootstrap samples, default is 1000.}

\item{conf.level}{Confidence level for calculating p-value thresholds, default is 0.95.}
}
\value{
A data frame with columns for each sample size pair (n1, n2) and the proportions of significant p-values for each test (ST, WT, NPBTT, WRST, PTT).
}
\description{
This function attempts to replicate and extend the simulation results
from Table 2 cell block 1.1 of the paper by Dwivedi et al. (2017).
The default parameter values aim to replicate the results from the paper,
while modifying the parameter values allows for an extension of the results.
}
\examples{
t2_replext_c1.1(n1 = c(4,5), n2 = c(4,5), n_simulations = 5)

}
\references{
Dwivedi AK, Mallawaarachchi I, Alvarado LA. Analysis of small sample size studies using nonparametric bootstrap test with pooled resampling method. Stat Med. 2017 Jun 30;36(14):2187-2205. doi: 10.1002/sim.7263. Epub 2017 Mar 9. PMID: 28276584.
}
